falta tight y splitNode

+++ src/node/RInternal.java
@@ -1,7 +1,7 @@
 package node;
 public class RInternal extends RNode{
 
-	public RInternal(int M, float m2) {
+	public RInternal(int M, int m2) {
 		super(M, m2);
 	}
