+++ src/node/RLeaf.java
@@ -1,7 +1,7 @@
 package node;
 public class RLeaf extends RNode{
 
-	public RLeaf(int M, float m2) {
+	public RLeaf(int M, int m2) {
 		super(M, m2);
 	}
